ctdbd does no longer take a --daemon parameter since we no longer do non-daemon mode remove the parameter from the ctdbd script
remove the store_unlock from ctdbd_test since there is no and will be no pdu for this
CTDB_REPLY_FETCH_LOCK no longer return the data for the record since the client is assumed to read this itself from the local tdb. remove some variables that no longer exists.