ctdb will now verify that the generation id for all CTDB_REQ_CALLs that it receives from remote nodes with its own generation number.
if the generation id do not match the daemon will just discard the 
packet completely.
this will require that clients such as samba3 will have to be aware that 
calls it sends to remote nodes may get lost and will have to resend the 
requests after a timeout.

during cluster recovery/reconfiguration the generation nnumbers will be 
changing and this can be used to prevent nodes that are still in the 
"old" instance from talking to nodes that belong to the new generation 
instance